+++ src/App.tsx
@@ -1,5 +1,5 @@
 import styled, { createGlobalStyle, css } from 'styled-components';
-import ToDoList from './ToDoList';
+import ToDoList from './components/ToDoList';
 
 const GlobalStyle = createGlobalStyle`
 ${css`
